1、修改 deleteOne() 和 delete()，返回值由 int 改为 int64。

+++ database/mongodb/callback.go
@@ -30,7 +30,7 @@ type BeforeDeleter interface {
 }
 
 type AfterDeleter interface {
-	AfterDelete(ctx context.Context, filter interface{}, opts []*options.DeleteOptions, count int) error
+	AfterDelete(ctx context.Context, filter interface{}, opts []*options.DeleteOptions, count int64) error
 }
 
 type BeforeFinder interface {
